splitting out task id things
